Release 0.2.0.
Rotated FCOS (anchor-free) with Hub 3× rIoU weights, decoded IoU loss, and TF/ONNX export.

+++ oriented_det/models/oriented_roi.py
@@ -412,8 +412,8 @@ def oriented_roi_align(
                         align_corners=False,
                     )
                 
-                # Store results in pre-allocated output
-                output_features[chunk_indices] = sampled_chunk
+                # grid_sample may promote to float32 under AMP; match feature-map dtype.
+                output_features[chunk_indices] = sampled_chunk.to(dtype=dtype)
     
     return output_features
